requirements.txt
ffmpeg-python==0.2.0
Pillow>=10.0.0
PySide6>=6.6.0
pytest>=8.0.0
pytest-cov>=4.1.0
pyinstaller>=6.0.0
